+++ chopper/codecov.yml
@@ -1,0 +1,4 @@
+coverage:
+  ignore:
+    - "lib/src/annotations.dart"
+    - "lib/src/constants.dart"
