+++ chopper/lib/chopper.dart
@@ -6,3 +6,4 @@ export 'src/interceptor.dart';
 export 'src/request.dart';
 export 'src/response.dart';
 export 'src/utils.dart';
+export 'src/constants.dart';
