+++ chopper/lib/src/constants.dart
@@ -1,0 +1,11 @@
+const contentTypeKey = 'content-type';
+const jsonHeaders = "application/json";
+const formEncodedHeaders = "application/x-www-form-urlencoded";
+
+class HttpMethod {
+  static const String Get = "GET";
+  static const String Post = "POST";
+  static const String Put = "PUT";
+  static const String Delete = "DELETE";
+  static const String Patch = "PATCH";
+}
