+++ chopper/pubspec.yaml
@@ -17,5 +17,6 @@ dev_dependencies:
   build_runner: ^1.0.0
   build_test:
   coverage: ^0.12.2
+  test_coverage: ^0.2.3
   chopper_generator:
     path: ../chopper_generator
